+++ src/NetCore/Web/Api/wwwroot/scripts/api/webgis.map.queryresults.js
@@ -1102,13 +1102,17 @@
                         if (featureMetadata) {
                             const linkname = $(this).data('linkname');
                             const linktarget = featureMetadata.linktargets[linkname];
-                            if (linktarget === 'dialog') {
-                                webgis.iFrameDialog(featureMetadata.links[linkname], linkname);
-                            }
-                            if (linktarget === 'datalinq_pdf_report') {
-                                webgis.downloadDataLinqPdf(featureMetadata.links[linkname]);
-                            } else {
-                                window.open(featureMetadata.links[linkname]);
+
+                            switch (linktarget) {
+                                case "dialog":
+                                    webgis.iFrameDialog(featureMetadata.links[linkname], linkname);
+                                    break;
+                                case "datalinq_pdf_report":
+                                    webgis.downloadDataLinqPdf(featureMetadata.links[linkname]);
+                                    break;
+                                default:
+                                    window.open(featureMetadata.links[linkname]);
+                                    break;
                             }
                         }
                     });